Add condition feature, and handling for monster and player 0hp
I have tried to make it simpler for now. You can now add conditions, but
om not too fan of the handling of it, but it works for now. Also
monsters will be removed from the row when hp is 0 or less. Still some
bugs that need fixing, and i should add a keymap to the table, and a
help page.